Minor bugs fixed.

+++ train_main.py
@@ -10,11 +10,12 @@ from models import create_model
 from utils import train_model, get_optimizer, get_loss, get_scheduler, test_model
 
 
-def train(_config, resume: bool = False):
+def train(_config, resume: bool = False, test: bool = False):
     print(json.dumps(config, indent=4))
 
     device = torch.device(_config['device'])
     os.environ["CUDA_VISIBLE_DEVICES"] = str(device.index)
+    device = torch.device(0)
 
     dataset = _config['data']['dataset']
     model_name = _config['model']['name']
@@ -37,22 +38,23 @@ def train(_config, resume: bool = False):
 
     save_folder = os.path.join('saves', dataset, _config['name'])
 
-    if not resume:
+    if not resume and not test:
         shutil.rmtree(save_folder, ignore_errors=True)
         os.makedirs(save_folder)
 
     with open(os.path.join(save_folder, 'config.yaml'), 'w+') as _f:
         yaml.safe_dump(_config, _f)
 
-    train_model(model=model,
-                dataset=dataset,
-                batch_size=_config['data']['batch-size'],
-                optimizer=optimizer,
-                scheduler=scheduler,
-                folder=save_folder,
-                trainer=trainer,
-                device=device,
-                **_config['train'])
+    if not test:
+        train_model(model=model,
+                    dataset=dataset,
+                    batch_size=_config['data']['batch-size'],
+                    optimizer=optimizer,
+                    scheduler=scheduler,
+                    folder=save_folder,
+                    trainer=trainer,
+                    device=device,
+                    **_config['train'])
 
     test_model(model=model,
                dataset=dataset,
@@ -68,6 +70,8 @@ if __name__ == '__main__':
                         help='Configuration filename for restoring the model.')
     parser.add_argument('--resume', required=False, type=bool, default=False,
                         help='Resume.')
+    parser.add_argument('--test', required=False, type=bool, default=False,
+                        help='Test.')
 
     args = parser.parse_args()
 
@@ -76,5 +80,8 @@ if __name__ == '__main__':
     if args.resume:
         print(f'Resume to {config["name"]}.')
         train(config, resume=True)
+    elif args.test:
+        print(f'Test {config["name"]}.')
+        train(config, test=True)
     else:
         train(config)
